Move ShaChain serializer to v1 package
