+++ src/commonMain/kotlin/fr/acinq/lightning/crypto/ShaChain.kt
@@ -2,25 +2,11 @@ package fr.acinq.lightning.crypto
 
 import fr.acinq.bitcoin.ByteVector32
 import fr.acinq.bitcoin.Crypto.sha256
-import fr.acinq.lightning.serialization.v1.ByteVector32KSerializer
 import fr.acinq.lightning.utils.startsWith
-import kotlinx.serialization.ExperimentalSerializationApi
-import kotlinx.serialization.KSerializer
-import kotlinx.serialization.Serializable
-import kotlinx.serialization.builtins.MapSerializer
-import kotlinx.serialization.builtins.serializer
-import kotlinx.serialization.descriptors.SerialDescriptor
-import kotlinx.serialization.descriptors.buildClassSerialDescriptor
-import kotlinx.serialization.descriptors.element
-import kotlinx.serialization.descriptors.mapSerialDescriptor
-import kotlinx.serialization.encoding.CompositeDecoder
-import kotlinx.serialization.encoding.Decoder
-import kotlinx.serialization.encoding.Encoder
 import kotlin.experimental.xor
 
 
 // TODO: should replace 0xffffffffffffffffL (-1) by 0xffffffffffffffL and 63 by 47.
-@Serializable(with = ShaChain.Serializer::class)
 data class ShaChain(val knownHashes: Map<List<Boolean>, ByteVector32>, val lastIndex: Long? = null) {
 
     fun addHash(hash: ByteVector32, index: Long): ShaChain {
@@ -101,47 +87,6 @@ data class ShaChain(val knownHashes: Map<List<Boolean>, ByteVector32>, val lastI
                     .minus<List<Boolean>, ByteVector32>(parentIndex + true)
                 addHash(receiver.copy(knownHashes = nodes1), hash, parentIndex)
             }
-
-    }
-
-    object Serializer : KSerializer<ShaChain> {
-        @OptIn(ExperimentalSerializationApi::class)
-        override val descriptor: SerialDescriptor = buildClassSerialDescriptor("ShaChain") {
-            element("knownHashes", mapSerialDescriptor(String.serializer().descriptor, ByteVector32KSerializer.descriptor))
-            element<Long>("lastIndex", isOptional = true)
-        }
-
-        private fun List<Boolean>.toBinaryString(): String = this.map { if (it) '1' else '0' }.joinToString(separator = "")
-        private fun String.toBooleanList(): List<Boolean> = this.map { it == '1' }
-
-        private val mapSerializer = MapSerializer(String.serializer(), ByteVector32KSerializer)
-
-        override fun serialize(encoder: Encoder, value: ShaChain) {
-            val compositeEncoder = encoder.beginStructure(descriptor)
-            compositeEncoder.encodeSerializableElement(descriptor, 0, mapSerializer, value.knownHashes.mapKeys { it.key.toBinaryString() })
-            if (value.lastIndex != null) compositeEncoder.encodeLongElement(descriptor, 1, value.lastIndex)
-            compositeEncoder.endStructure(descriptor)
-        }
-
-        override fun deserialize(decoder: Decoder): ShaChain {
-            var knownHashes: Map<List<Boolean>, ByteVector32>? = null
-            var lastIndex: Long? = null
-
-            val compositeDecoder = decoder.beginStructure(descriptor)
-            loop@ while (true) {
-                when (compositeDecoder.decodeElementIndex(descriptor)) {
-                    CompositeDecoder.DECODE_DONE -> break@loop
-                    0 -> knownHashes = compositeDecoder.decodeSerializableElement(descriptor, 0, mapSerializer).mapKeys { it.key.toBooleanList() }
-                    1 -> lastIndex = compositeDecoder.decodeLongElement(descriptor, 1)
-                }
-            }
-            compositeDecoder.endStructure(descriptor)
-
-            return ShaChain(
-                knownHashes ?: error("No knownHashes in structure"),
-                lastIndex
-            )
-        }
 
     }
 }
